ecs: widen Health to f64 — f32 made damage numbers garbage on big mobs
All three client-verification lenses independently found this, and one
reproduced it against the real config numbers. It is my bug, from the original
component catalog: I reasoned carefully about f64 for Position (the maze sits
at 200000,200000) and never applied the same test to Health.

Mob health is base x HEALTH_SCALING, which leaves f32's exact-integer range
entirely:

  target_dummy/apex   3.0e15 HP   f32 ulp 2.8e7
  digger/apex         3.0e10 HP   f32 ulp 4.1e3
  ant_hole/unique     5.0e9  HP   exact

The client computes its floating damage number as previousHealth - newHealth.
With health stored f32, a 500-damage hit on an apex target dummy lands entirely
inside one ulp: the stored value does not move, so the readout is wrong and
sometimes ZERO. A verifier simulated five identical 500-damage hits on an apex
digger and got "-1524", then nothing at all, then "-476", "-976", "-1476".

The target dummy is the one entity in the game whose entire purpose is
displaying damage numbers, so this is not a rounding curiosity. It also fed
graphics.getAccumulatedDamage, so the on-kill total was wrong too.

Damage.value deliberately STAYS f32: real mob damage tops out near 5e5, well
inside f32's exact range, and there is no accumulation on that field.

Verified against the exact failing case: 3e15 HP, one 500-damage hit, stored
delta is now exactly 500. Gates all pass; harness unchanged.

+++ src/ecs/components/combat.ts
@@ -9,10 +9,23 @@
 
 import { defineComponent, defineTag } from '../component';
 
-/** Current and maximum hit points. */
+/**
+ * Current and maximum hit points.
+ *
+ * f64, NOT f32, for the same reason Position is — the values are far outside
+ * f32's exact-integer range. Mob health is base x HEALTH_SCALING, and an apex
+ * target dummy is 3e15 HP where one f32 ulp is 2.8e7. A 500-damage hit lands
+ * entirely inside the rounding error: the stored value does not change, so the
+ * client's damage readout (previousHealth - newHealth) is arithmetic garbage
+ * and sometimes zero. An apex digger at 3e10 has a 4096 ulp and drifts too.
+ *
+ * The target dummy is the one entity whose entire purpose is displaying damage
+ * numbers, so this is not a rounding curiosity. Damage.value stays f32: real
+ * mob damage tops out around 5e5, well inside f32's exact range.
+ */
 export const Health = defineComponent('Health', {
-    current: 'f32',
-    max: 'f32',
+    current: 'f64',
+    max: 'f64',
 });
 
 /** Contact/attack damage this entity deals. */
